+++ src/main/java/com/stu/mapper/EvaluationMapper.java
@@ -10,13 +10,13 @@ import java.util.List;
 @Mapper
 public interface EvaluationMapper {
     @Select("select * from evaluation where student_id = #{studentId} and teacher_id = #{teacherId}")
-    Evaluation getByStudentIdAndTeacherId(Long studentId, Integer teacherId);
+    Evaluation getByStudentIdAndTeacherId(Long studentId, Long teacherId);
 
     @Insert("insert into evaluation (student_id, teacher_id, score, context) values (#{studentId}, #{teacherId}, #{score}, #{context})")
     void insert(Evaluation evaluation);
 
     @Select("select * from evaluation where teacher_id = #{teacherId}")
-    List<Evaluation> getByTeacherId(Integer teacherId);
+    List<Evaluation> getByTeacherId(Long teacherId);
 
     @Select("select * from evaluation where student_id = #{currentId}")
     List<Evaluation> getByUserId(Long currentId);
